change total amount type to big decimal

+++ src/main/java/com/api/zorvanz/domain/cart/CartData.java
@@ -2,11 +2,12 @@ package com.api.zorvanz.domain.cart;
 
 import com.api.zorvanz.domain.cartitem.CartItemData;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartData(
         Long id,
-        Double totalAmount,
+        BigDecimal totalAmount,
         List <CartItemData> cartItems
 ) {
 }
